+++ lib/dropdown_custom.dart
@@ -13,6 +13,7 @@ export 'src/custom_dropdown.dart'
         ItemEnabled,
         DropdownItemBuilder,
         DropdownErrorBuilder;
+export 'src/dropdown_chip_style.dart' show ChipOverflow, DropdownChipStyle;
 export 'src/dropdown_direction.dart' show DropdownDirection;
 export 'src/dropdown_field_style.dart' show DropdownFieldStyle;
 export 'src/dropdown_loading.dart' show DropdownLoading;